fix bug in `git init --separate-git-dir`
the dir specified with --separate-git-dir is the *actual* place to
create the git repo files i.e. it *is*
literally the '.git' dir. It's not the parent of the .git dir, whereas
the directory given to `git init` is the parent e.g. `git init foo`
means create foo/.git.