fix hash-object -w option and refactor (#12)
* initial commit of project generated by 'gradle init'

* fixup project structure

* implement fortune telling

* change to the "executable" is called 'fortune'

* fix perm on gradlew

* Initialize Rust-based Fortune application

Added configuration file Cargo.toml, defining the new package "fortune-rust" and its dependencies. Created main.rs to implement the core functionality, including reading and selecting a random fortune from files.

* initial commit with partial support for git init, hash-object, cat-file

* enhance hash-object command and refactor cat-file

support the -w (write) option for hash-object. Refactor the cat-file
code to use the full cmdline args

* support executing commands from subdir of git repo

Git allows running commands in any directory within the git repo tree.
So add support to find the .git dir in a parent or panic if not found.

* adjust cli options

* partial obj_id matching for cat-file

add support to find the correct object file with a shortened obj hash
e.g. cat-file -p 1234

* enhance cat-file command

refactor to support -t and -s options

* support cat-file -p for tree objects

* enhance 'init' command support

* read ~/.gitconfig to use init.defaultBranch when writing .git/HEAD
* use --initial-branch if present
* write .git/config with proper values

* enhance 'init' command support

* read ~/.gitconfig to use init.defaultBranch when writing .git/HEAD
* use --initial-branch if present
* write .git/config with proper values

* init command enhancement for alternate git dirs

support the --separate-git-dir option as well as the directory if
specified i.e. `git init foo` and
`git init --separate-git-dir foo bar`

* add initial 'git config' command support

support -l/--list and refactor the section matching to use regex to
simplify the code

* refactor how the initial branch is determined

clean up the handling of the init.defaultBranch vs --initial-branch and
the fallback to 'master' as the branch name

* Create README.md

* fix readme formatting

* fix bug in `git init --separate-git-dir`

the dir specified with --separate-git-dir is the *actual* place to
create the git repo files i.e. it *is*
literally the '.git' dir. It's not the parent of the .git dir, whereas
the directory given to `git init` is the parent e.g. `git init foo`
means create foo/.git.

* fix broken `hash-object -w`

the prior changes to the constants with the various directory names
broke the write option of hash-object so fix that and refactor to clean
up for readability