Git write tree (#18)
* begin work on ls-tree command

* basic ls-tree impl

refactor to create a GitObject and related module to simplify getting
objects, then use that in both cat-file and ls-tree. The function in
cat-file that was printing tree objects is moved to the ls-tree and used
from cat-file.

* more ls-tree support

support some of the cli args for what to print and separate tags out
into its own module

* fix hash object hashing

* fix hash object

* handle -r option to recursively print tree objects

* only call create_file_name() once

* support ls-tree via a commit object

ls-tree works when given a commit hash either directly or via a tag

* tags are not objects

* wip with write-tree

* additional write-tree work

make sure the mode is saved correctly and the entries are sorted

* implement commit-tree

mostly complete except need to have full sha1 hashes

* finalize commit-tree

make sure the sha1 hashes are full hashes

* fix copy/paste error in commit_tree.rs

* Refactor struct visibility for encapsulation

Changed the visibility of several struct fields to private in `Commit`
and `Tag` to enforce encapsulation. Removed an unnecessary default match
 arm in `cat_file.rs` for cleaner code.