Push wnznvmnmqkxr (#23)
* new blog example for type state pattern

initial commit of shell generated by 'cargo init'

* write typestate example

* cleanup

* flesh out the type state example

* enforce that build() can only be called once source is provided

use the type state pattern in the builder as well, to ensure the
only way to call CompilerBuilder.build() is if CompilerBuilder.source()
has been called.

* use the builder provided configuration

optionally print the tokens and ast

* simplify by combining impl blocks

* cleanup slightly but in parituclary the build() method

Prior to implementing the state pattern for the builder, we used a
Result object with an if test in the method of the body, but one we did
implement the state pattern, I forgot to remove that.

Also simplified where I could by using `..self`

* very minor cleanup

* split into lib and bin 

separate the compiler and builder into a separate module to validate
that at each stage we can do only what we're supposed to, adjusting
visibility appropriately.